TasksAPI, Ormar & Pydantic, ResetPasswordUsers, AccessSecutiry
* Added a new api section for working with tasks.
* Adjustments have been made to the database and Pydantic models.
* Added deletion of outdated data from the ResetPasswordUsers table.
* Added additional checks for access to api resources.

+++ backend/projects/models.py
@@ -7,6 +7,6 @@ from users.models import User
 class Project(ormar.Model):
     ormar_config = base_ormar_config.copy(tablename="projects")
 
-    id: int = ormar.Integer(primary_key=True)
+    id: int = ormar.BigInteger(primary_key=True)
     title: str = ormar.String(nullable=False, max_length=50)
     creator: User = ormar.ForeignKey(User, nullable=False, ondelete=ormar.ReferentialAction.CASCADE)
